Move ReactNativeTypes-flowtest.js out of the Renderer folder
Summary:
The renderer folder gets replaced by react syncs but we don't want to lose this flow test. It would make sense to move this file into the React repo as it is trying to test `ReactNativeTypes.js` which lives there. However, the React repo is on Flow 0.72 which apparently doesn't catch this issue.

Changelog: [Internal]

(Note: this ignores all push blocking failures!)

Reviewed By: gaearon

Differential Revision: D18231354

fbshipit-source-id: 69f4617899c06118c74a8991f061e3bb3fdc88fb