Added @matt-oakes as CODEOWNER of the ESLint package (#27689)
Summary:
Adding myself as a code owner for the files in the ESLint package so I get email alerts for all PRs which alter the package.

This is needed because I need to manually publish updates to the package when pull requests are merged in. They are not handled by the standard React Native release cycle.

Related discussion: https://github.com/facebook/react-native/pull/27085#issuecomment-571089097

## Changelog

[Internal] [Added] - Added mat-oakes as a CODEOWNER for the ESLint package
Pull Request resolved: https://github.com/facebook/react-native/pull/27689

Test Plan: N/A

Differential Revision: D19294768

Pulled By: cpojer

fbshipit-source-id: 3e3833261fc0b3c107b52b0e8232bd942a2ecaf4